move Paged to models and add unit tests
